% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/british_isles.R
\docType{data}
\name{british_isles}
\alias{british_isles}
\title{British Isles}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 1 columns.
}
\source{
The rnaturalearth package
}
\usage{
british_isles
}
\description{
A dataset containing the British Isles outline map
}
\keyword{datasets}
