% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpat_gridcreate.R
\name{gpat_gridcreate}
\alias{gpat_gridcreate}
\title{Grid polygon creator}
\usage{
gpat_gridcreate(x, proj = TRUE)
}
\arguments{
\item{x}{A filepath to the geoPAT 2.0 grid header file}

\item{proj}{TRUE/FALSE; should a new grid polygon inherit spatial projection from the geoPAT 2.0 grid header file}
}
\value{
sfc_POLYGON
}
\description{
Creates a polygon of the gpat grid based on a grid header
}
\examples{
header_filepath = system.file("rawdata/Augusta2011_grid100.hdr", package="rgeopat2")
my_grid = gpat_gridcreate(header_filepath)

}
