% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpat_create_grid.R
\name{gpat_header_parser}
\alias{gpat_header_parser}
\title{Parse a header of a GeoPAT 2 grid file}
\usage{
gpat_header_parser(x)
}
\arguments{
\item{x}{A filepath to the GeoPAT 2 grid header file}
}
\value{
data_frame
}
\description{
Extracts basic information from a geoPAT 2 grid header file
}
