% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpat_read_distmtx.R
\name{gpat_read_distmtx}
\alias{gpat_read_distmtx}
\title{Read a GeoPAT distance matrix}
\usage{
gpat_read_distmtx(x)
}
\arguments{
\item{x}{A filepath to the geoPAT 2 distance matrix file}
}
\value{
dist
}
\description{
Read a GeoPAT distance matrix into R
}
\examples{
distmtx_filepath = system.file("rawdata/Augusta2011_matrix_grid.csv", package="rgeopat2")
my_distmtx = gpat_read_distmtx(distmtx_filepath)

}
