% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeopat2-package.R
\docType{package}
\name{rgeopat2-package}
\alias{rgeopat2}
\alias{rgeopat2-package}
\title{rgeopat2: Additional Functions for 'GeoPAT' 2}
\description{
Supports analysis of spatial data processed with the 'GeoPAT' 2 software \url{https://github.com/Nowosad/geopat2}. Available features include creation of a grid based on the 'GeoPAT' 2 grid header file and reading a 'GeoPAT' 2 text outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Nowosad/rgeopat2}
  \item Report bugs at \url{https://github.com/Nowosad/rgeopat2/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

Other contributors:
\itemize{
  \item Space Informatics Lab [copyright holder]
}

}
\keyword{internal}
