% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpat_gridcreate.R
\name{gpat_header_parser}
\alias{gpat_header_parser}
\title{Parse a header of a geoPAT 2.0 grid file}
\usage{
gpat_header_parser(x)
}
\arguments{
\item{x}{A filepath to the geoPAT 2.0 grid header file}
}
\value{
data_frame
}
\description{
Extracts basic information from a geoPAT 2.0 grid header file
}
